+++ COMP313003_202101_Project/Areas/Orders/Controllers/Api/PurchaseOrderLineController.cs
@@ -120,7 +120,7 @@ namespace COMP313003_202101_Project.Areas.Orders.Controllers.Api
         public IActionResult Remove([FromBody]CrudViewModel<PurchaseOrderLine> payload)
         {
             PurchaseOrderLine purchaseOrderLine = _context.PurchaseOrderLine
-                .Where(x => x.PurchaseOrderLineId == (int)payload.key)
+                .Where(x => x.PurchaseOrderLineId == (int)(long)payload.key)
                 .FirstOrDefault();
             _context.PurchaseOrderLine.Remove(purchaseOrderLine);
             _context.SaveChanges();
